Rename vite-server and build IPC methods
